// src/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Yeying API
 * 
 *
 * OpenAPI spec version: 1.0
 * Contact: yeying.community@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import * as application from './api/application/types'
import * as archive from './api/archive/types'
import * as asset from './api/asset/types'
import * as assignment from './api/assignment/types'
import * as audit from './api/audit/types'
import * as block from './api/block/types'
import * as bulletin from './api/bulletin/types'
import * as certificate from './api/certificate/types'
import * as config from './api/config/types'
import * as content from './api/content/types'
import * as context from './api/context/types'
import * as event from './api/event/types'
import * as experience from './api/experience/types'
import * as group from './api/group/types'
import * as homework from './api/homework/types'
import * as identity from './api/identity/types'
import * as invitation from './api/invitation/types'
import * as knowledge from './api/knowledge/types'
import * as link from './api/link/types'
import * as llm from './api/llm/types'
import * as mail from './api/mail/types'
import * as message from './api/message/types'
import * as minio from './api/minio/types'
import * as mistakes from './api/mistakes/types'
import * as namespace from './api/namespace/types'
import * as network from './api/network/types'
import * as node from './api/node/types'
import * as provider from './api/provider/types'
import * as recycle from './api/recycle/types'
import * as room from './api/room/types'
import * as service from './api/service/types'
import * as session from './api/session/types'
import * as social from './api/social/types'
import * as spider from './api/spider/types'
import * as support from './api/support/types'
import * as task from './api/task/types'
import * as taskTag from './api/taskTag/types'
import * as topic from './api/topic/types'
import * as user from './api/user/types'
import * as vector from './api/vector/types'
import * as wallet from './api/wallet/types'
import * as warehouse from './api/warehouse/types'

export interface ApiImplementation {
	application: application.ApplicationApi
	archive: archive.ArchiveApi
	asset: asset.AssetApi
	assignment: assignment.AssignmentApi
	audit: audit.AuditApi
	block: block.BlockApi
	bulletin: bulletin.BulletinApi
	certificate: certificate.CertificateApi
	config: config.ConfigApi
	content: content.ContentApi
	context: context.ContextApi
	event: event.EventApi
	experience: experience.ExperienceApi
	group: group.GroupApi
	homework: homework.HomeworkApi
	identity: identity.IdentityApi
	invitation: invitation.InvitationApi
	knowledge: knowledge.KnowledgeApi
	link: link.LinkApi
	llm: llm.LlmApi
	mail: mail.MailApi
	message: message.MessageApi
	minio: minio.MinioApi
	mistakes: mistakes.MistakesApi
	namespace: namespace.NamespaceApi
	network: network.NetworkApi
	node: node.NodeApi
	provider: provider.ProviderApi
	recycle: recycle.RecycleApi
	room: room.RoomApi
	service: service.ServiceApi
	session: session.SessionApi
	social: social.SocialApi
	spider: spider.SpiderApi
	support: support.SupportApi
	task: task.TaskApi
	taskTag: taskTag.TaskTagApi
	topic: topic.TopicApi
	user: user.UserApi
	vector: vector.VectorApi
	wallet: wallet.WalletApi
	warehouse: warehouse.WarehouseApi
}

export interface ApiOptions {
	failOnUnknownProperties?: boolean
}